requirements.txt
streamlit==1.27.0
ultralytics==8.0.134
opencv-python
numpy==1.26.4
matplotlib
Pillow
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
